+++ packages/react/src/icons/ToggleRight.tsx
@@ -1,0 +1,3 @@
+import * as React from "react";
+import type { SVGProps } from "react";
+export const ToggleRight = (props: SVGProps<SVGSVGElement>) => <svg width="1em" height="1em" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}><rect x={4} y={7} width={16} height={10} rx={5} stroke="black" strokeWidth={2} /><circle cx={15} cy={12} r={2} fill="black" stroke="black" strokeWidth={2} /></svg>;