+++ packages/svelte/src/icons/ToggleRight.svelte
@@ -1,0 +1,23 @@
+<script lang="ts">
+  export let size: number | string = 24;
+  export let color: string = 'currentColor';
+  export let strokeWidth: number | string = 2;
+  export let width: number | string = size;
+  export let height: number | string = size;
+</script>
+
+<svg
+  xmlns="http://www.w3.org/2000/svg"
+  {width}
+  {height}
+  viewBox="0 0 24 24"
+  fill="none"
+  stroke={color}
+  stroke-width={strokeWidth}
+  stroke-linecap="round"
+  stroke-linejoin="round"
+  {...$$restProps}
+>
+  <rect x="4" y="7" width="16" height="10" rx="5" stroke="black" stroke-width="2"/>
+<circle cx="15" cy="12" r="2" fill="black" stroke="black" stroke-width="2"/>
+</svg>